Refactored vle and preview module initialization code into a function. #2424
